requirements.txt
astroid==2.5.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
gunicorn==20.1.0
idna==2.10
isort==5.8.0
itsdangerous==1.1.0
Jinja2==2.11.3
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
mccabe==0.6.1
pkg-resources==0.0.0
pushbullet.py==0.12.0
pylint==2.8.2
python-magic==0.4.22
requests==2.25.1
rpi==0.0.1
RPi.GPIO==0.7.0
toml==0.10.2
typed-ast==1.4.3
urllib3==1.26.5
websocket-client==1.0.1
Werkzeug==1.0.1
wrapt==1.12.1
